research_suite/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stegresearch"
version = "0.1.0"
description = "Modular steganography research suite for lawful analysis and watermarking"
requires-python = ">=3.11"
authors = [{name = "Research Team"}]
readme = "README.md"
dependencies = [
  "numpy>=1.26",
  "scipy>=1.11",
  "pillow>=10.2",
  "opencv-python>=4.9",
  "PyWavelets>=1.5",
  "scikit-image>=0.22",
  "scikit-learn>=1.4",
  "joblib>=1.3",
  "pydub>=0.25",
  "soundfile>=0.12",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "click>=8.1",
  "rich>=13.7",
  "pydantic>=2.6",
  "python-multipart>=0.0.9",
  "pydantic-settings>=2.6",
  "jinja2>=3.1",
  "httpx>=0.27",
  "structlog>=24.1",
  "pyyaml>=6.0",
  "scapy>=2.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.1",
  "hypothesis>=6.99",
  "mypy>=1.8",
  "ruff>=0.3",
  "black>=24.3",
  "types-Pillow",
  "types-setuptools",
]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B", "UP", "S", "N", "W"]
ignore = ["S101", "B008"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = []

[tool.pytest.ini_options]
addopts = "-ra --strict-markers --strict-config --maxfail=1"
render_collapsed = true
testpaths = ["tests"]
filterwarnings = ["error"]

[tool.coverage.run]
source = ["stegresearch"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true
